feat: 면접 생성 및 진행 API

+++ src/main/kotlin/com/tik/server/controller/ResumeController.kt
@@ -27,7 +27,7 @@ class ResumeController(
 
     @DeleteMapping("/delete/{resumeId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    fun deleteResume(@PathVariable resumeId: Long) {
+    fun deleteResume(@PathVariable resumeId: Int) {
         resumeService.deleteResume(resumeId)
     }
