+++ src/main/kotlin/com/tik/server/repository/ProjectTechStackRepository.kt
@@ -3,5 +3,5 @@ package com.tik.server.repository
 import com.tik.server.entity.ProjectTechStack
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ProjectTechStackRepository : JpaRepository<ProjectTechStack, Long> {
+interface ProjectTechStackRepository : JpaRepository<ProjectTechStack, Int> {
 }
